modularize pwm sine wave generation
