+++ pwm.c
@@ -1,0 +1,71 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <math.h>
+
+#include "pico/stdlib.h"   
+#include "pico/multicore.h"
+#include "hardware/irq.h"  
+#include "hardware/pwm.h"  
+#include "hardware/sync.h" 
+#include "hardware/gpio.h"
+
+#include "pwm.h"
+#include "sine.h"
+
+// initialize sine table index
+int sine_position = 0;
+
+/*
+ * PWM Interrupt Handler which outputs PWM level and advances the 
+ * current sample. 
+ * 
+ * We repeat the same value for 2^SAMPLE_RATE_FACTOR_SHIFT cycles. This means 
+ * the sample rate is adjusted by factor of 2^SAMPLE_RATE_FACTOR_SHIFT.
+ */
+void pwm_interrupt_handler() {
+    // clear interrupt flag
+    pwm_clear_irq(pwm_gpio_to_slice_num(INPUT_SIGNAL_PIN));    
+
+    if (sine_position < (SINE_TABLE_LENGTH<<SAMPLE_RATE_FACTOR_SHIFT) - 1) { 
+        // set pwm level 
+        pwm_set_gpio_level(INPUT_SIGNAL_PIN, SINE_TABLE[sine_position>>SAMPLE_RATE_FACTOR_SHIFT]);  
+        sine_position++;
+    } else {
+        // reset to start
+        sine_position = 0;
+    }
+}
+
+/*
+ * code that core 1 will run
+ *
+ * generates a sine wave using pwm
+ */
+void generate_sine_wave(uint32_t sine_frequency) {
+    // initialize pwm pin
+    gpio_set_function(INPUT_SIGNAL_PIN, GPIO_FUNC_PWM);
+    
+    int pin_slice = pwm_gpio_to_slice_num(INPUT_SIGNAL_PIN);
+
+    // Setup PWM interrupt to fire when PWM cycle is complete
+    pwm_clear_irq(pin_slice);
+    pwm_set_irq_enabled(pin_slice, true);
+    // use interrupt handler function defined above
+    irq_set_exclusive_handler(PWM_IRQ_WRAP, pwm_interrupt_handler); 
+    irq_set_enabled(PWM_IRQ_WRAP, true);
+
+    // initialize pwm config: clock division and wrap
+    pwm_config config = pwm_get_default_config();
+    float clkdiv = CLK_KHZ * KHZ_TO_HZ / (WRAP * (SINE_TABLE_LENGTH<<SAMPLE_RATE_FACTOR_SHIFT) * sine_frequency);
+    pwm_config_set_clkdiv(&config, clkdiv); 
+    pwm_config_set_wrap(&config, WRAP); 
+    pwm_init(pin_slice, &config, true);
+
+    // set initial pwm level
+    pwm_set_gpio_level(INPUT_SIGNAL_PIN, 0);
+
+    while(1) {
+        __wfi(); // Wait for Interrupt
+    }
+}
